feat(Diets): Added Diets Controller.

+++ RunningGo.API/Dietas/Domain/Models/Diet.cs
@@ -15,6 +15,6 @@ public class Diet
     public int FoodId { set; get; }
     public Food Food;
     
-    public int UserId { set; get; }
+    public long UserId { set; get; }
     public User User;
 }
